FACING THE FIRST CHALLENGE
When changing the option of the drop down menu, the previous option remains. This will be confusing for the end user and takes up a lot of real estate. So to fix this, there are two solutions: have the user select what they want to change on the previous screen and lead to a new screen that is solely dedicated to changing that value. OR, finding a method to constantly update the screen with the changes (Change the value before the screen is displayed)