Code optimisation for managing account (NOT TESTED)
Previous, had it so that for changing any details, there would be a separate function for each (password, address1, address2 etc), when they all basically did the same thing. Now have it so that they would all be done within one function. As the success criteria had said no unnecessary lines of repeated code.
NOT TESTED
The only other portions of repeated code are the portions for searching through each other category but this CANNOT be condensed